requirements.txt
torch>=1.8.0
torchvision
opencv-python
pyyaml
matplotlib
pandas
requests
